+++ .pre-commit-config.yaml
@@ -13,6 +13,8 @@ repos:
     rev: 1.13.0
     hooks:
     -   id: blacken-docs
+        additional_dependencies:
+          - black==23.3.0
   - repo: https://github.com/codespell-project/codespell
     rev: v2.2.4
     hooks:
