+++ .readthedocs.yml
@@ -1,0 +1,18 @@
+version: 2
+build:
+  os: ubuntu-22.04
+
+  tools:
+    python: "3.10"
+
+  apt_packages:
+    - libpango1.0-dev
+    - ffmpeg
+    - graphviz
+
+python:
+   install:
+      - requirements: docs/rtd-requirements.txt
+      - requirements: docs/requirements.txt
+      - method: pip
+        path: .
