+++ docs/Makefile
@@ -1,0 +1,33 @@
+# Minimal makefile for Sphinx documentation
+#
+
+# You can set these variables from the command line, and also
+# from the environment for the first two.
+SPHINXOPTS    ?=
+SPHINXBUILD   ?= sphinx-build
+
+# Path base is the source directory
+SOURCEDIR     = .
+BUILDDIR      = ../build
+
+# Put it first so that "make" without argument is like "make help".
+help:
+	@(cd source; $(SPHINXBUILD) -M help "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O))
+
+.PHONY: help Makefile i18n
+
+# All the code is executed as if everything was launched in one shell.
+.ONESHELL:
+# Like make clean but also remove files generated by autosummary and
+# rendered videos.
+cleanall: clean
+	@rm source/reference/*
+	@rm -rf source/media
+	@rm -f rendering_times.csv
+i18n:
+	@(cd source; $(SPHINXBUILD) -M gettext "$(SOURCEDIR)" ../i18n/ -t skip-manim $(SPHINXOPTS) $(O);cd ../i18n;bash stripUntranslatable.sh)
+
+# Catch-all target: route all unknown targets to Sphinx using the new
+# "make mode" option.  $(O) is meant as a shortcut for $(SPHINXOPTS).
+%: Makefile
+	@(cd source; $(SPHINXBUILD) -M $@ "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O))
