+++ docs/requirements.txt
@@ -1,0 +1,5 @@
+furo
+myst-parser
+sphinx<5.1
+sphinx-copybutton
+sphinxext-opengraph
