+++ docs/rtd-requirements.txt
@@ -1,0 +1,2 @@
+jupyterlab
+sphinxcontrib-programoutput
