+++ docs/source/_static/custom.css
@@ -1,0 +1,84 @@
+@media (prefers-color-scheme: dark) {
+    span.nc {
+        text-decoration: none !important;
+    }
+}
+
+.admonition-manim-example {
+    padding: 0;
+    display: flex;
+    flex-direction: column;
+}
+
+.admonition-manim-example p.admonition-title {
+    font-weight: 600;
+    font-size: 0.925rem;
+    margin: 0;
+}
+
+.admonition-manim-example .highlight-python {
+    margin: 0;
+}
+
+.admonition-manim-example .highlight {
+    border-radius: 0;
+}
+
+.admonition-manim-example .highlight pre {
+    font-size: 15px;
+}
+
+.manim-video {
+    width: 100%;
+    padding: 8px 0;
+    outline: 0;
+}
+
+.admonition-manim-example .manim-video {
+    padding: 0;
+}
+
+.admonition-manim-example img {
+    margin-bottom: 0;
+}
+
+.admonition-manim-example p:last-child {
+    margin-top: 0;
+    padding-left: 0.5rem;
+    padding-bottom: 0.15rem;
+    font-size: 15px;
+}
+
+.admonition-manim-example .copybtn {
+    margin-right: 6px;
+    font-size: 18px;
+}
+
+.admonition-manim-example .copybtn:hover {
+    cursor: pointer;
+}
+
+p.rubric{
+    text-transform: capitalize;
+    font-size: 1.25rem;
+    font-weight: bold;
+}
+
+.sig-param{
+    color: var(--color-content-foreground);
+}
+
+dl.c .field-list dt, dl.cpp .field-list dt, dl.js .field-list dt, dl.py .field-list dt {
+    text-transform: capitalize;
+    font-weight: bold;
+    font-size: var(--font-size--normal);
+}
+
+h4, h5, h6{
+    text-transform: none;
+}
+
+/* yikes-ish attempt at bugfix for navbar on some browsers */
+.sidebar-tree a.internal.reference {
+    display: table-cell;
+}
