+++ docs/source/_templates/autosummary/class.rst
@@ -1,0 +1,32 @@
+{{ name | escape | underline}}
+
+Qualified name: ``{{ fullname | escape }}``
+
+.. currentmodule:: {{ module }}
+
+.. autoclass:: {{ objname }}
+   :show-inheritance:
+   :members:
+
+   {% block methods %}
+   {%- if methods %}
+   .. rubric:: {{ _('Methods') }}
+
+   .. autosummary::
+      :nosignatures:
+      {% for item in methods if item != '__init__' and item not in inherited_members %}
+      ~{{ name }}.{{ item }}
+      {%- endfor %}
+   {%- endif %}
+   {%- endblock %}
+
+   {% block attributes %}
+   {%- if attributes %}
+   .. rubric:: {{ _('Attributes') }}
+
+   .. autosummary::
+     {% for item in attributes %}
+     ~{{ name }}.{{ item }}
+     {%- endfor %}
+   {%- endif %}
+   {% endblock %}
