+++ docs/source/index.rst
@@ -1,0 +1,15 @@
+Manim Physics
+-------------
+
+Welcome to the `manim-physics` official documentation!
+
+Index
++++++
+
+.. toctree::
+   :maxdepth: 3
+
+   reference_index/electromagnetism
+   reference_index/optics
+   reference_index/rigid_mechanics
+   reference_index/wave
