+++ docs/source/reference_index/electromagnetism.rst
@@ -1,0 +1,10 @@
+Electromagnetism
+========
+
+.. currentmodule:: manim_physics
+
+.. autosummary::
+   :toctree: ../reference
+
+   ~electromagnetism.electrostatics
+   ~electromagnetism.magnetostatics
