+++ docs/source/reference_index/optics.rst
@@ -1,0 +1,12 @@
+Optics
+------
+
+
+.. currentmodule:: manim_physics
+
+.. autosummary::
+   :toctree: ../reference
+
+    ~optics.lenses
+    ~optics.rays
+
