+++ docs/source/reference_index/rigid_mechanics.rst
@@ -1,0 +1,10 @@
+Rigid Mechanics
+---------------
+
+.. currentmodule:: manim_physics
+
+.. autosummary::
+   :toctree: ../reference
+
+   ~rigid_mechanics.rigid_mechanics
+   ~rigid_mechanics.pendulum
