+++ docs/source/reference_index/wave.rst
@@ -1,0 +1,10 @@
+Waves
+========
+
+.. currentmodule:: manim_physics
+
+.. autosummary::
+   :toctree: ../reference
+
+   ~wave
+
