+++ manim_physics/__init__.py
@@ -4,7 +4,8 @@ from manim import *
 
 from .electromagnetism.electrostatics import *
 from .electromagnetism.magnetostatics import *
-from .lensing import *
-from .pendulum import *
-from .rigid_mechanics import *
+from .optics.lenses import *
+from .optics.rays import *
+from .rigid_mechanics.pendulum import *
+from .rigid_mechanics.rigid_mechanics import *
 from .wave import *
