+++ manim_physics/electromagnetism/magnetostatics.py
@@ -13,6 +13,28 @@ import numpy as np
 
 
 class Wire(VMobject, metaclass=ConvertToOpenGL):
+    """An abstract class denoting a current carrying wire to produce a
+    :class:`~MagneticField`.
+
+    Parameters
+    ----------
+    stroke
+        The original wire ``VMobject``. The resulting wire takes its form.
+    current
+        The magnitude of current flowing in the wire.
+    samples
+        The number of segments of the wire used to create the
+        :class:`~MagneticField`.
+    kwargs
+        Additional parameters passed to ``VMobject``.
+
+
+    .. note::
+
+        See :class:`~MagneticField` for examples.
+
+    """
+
     def __init__(
         self,
         stroke: VMobject,
@@ -28,6 +50,35 @@ class Wire(VMobject, metaclass=ConvertToOpenGL):
 
 
 class MagneticField(ArrowVectorField):
+    """A magnetic field.
+
+    Parameters
+    ----------
+    wires
+        All wires contributing to the total field.
+    kwargs
+        Additional parameters to be passed to ``ArrowVectorField``.
+
+    Example
+    -------
+    .. manim:: MagneticFieldExample
+        :save_last_frame:
+
+        from manim_physics import *
+
+        class MagneticFieldExample(ThreeDScene):
+            def construct(self):
+                wire = Wire(Circle(2).rotate(PI / 2, UP))
+                mag_field = MagneticField(
+                    wire,
+                    x_range=[-4, 4],
+                    y_range=[-4, 4],
+                )
+                self.set_camera_orientation(PI / 3, PI / 4)
+                self.add(wire, mag_field)
+
+    """
+
     def __init__(self, *wires: Wire, **kwargs):
         dls = []
         currents = []
