+++ manim_physics/optics/__init__.py
@@ -1,0 +1,5 @@
+"""A lensing module.
+
+Currently only shows refraction in lenses and not
+total internal reflection.
+"""
