+++ pyproject.toml
@@ -12,10 +12,18 @@ manim = "^0.17.3"
 pymunk = "^6.0.0"
 Shapely = "^1.8.0"
 
+[tool.poetry.group.dev]
+optional = true
+
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.3.1"
 black = "^23.3.0"
 pre-commit = "^3.3.2"
+furo = "^2022.06.21"
+sphinx = "^4"
+sphinx-copybutton = "^0.4.0"
+sphinxcontrib-programoutput = "^0.17"
+myst-parser = "^0.17.2"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
